move ImageRemovalCheck type definition to types.go
